+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/modelselection/metrics/ClusteringMetrics.java
@@ -17,7 +17,7 @@ package com.datumbox.framework.core.machinelearning.modelselection.metrics;
 
 import com.datumbox.framework.common.dataobjects.Dataframe;
 import com.datumbox.framework.common.dataobjects.Record;
-import com.datumbox.framework.core.machinelearning.common.abstracts.validators.AbstractMetrics;
+import com.datumbox.framework.core.machinelearning.common.abstracts.modelselection.metrics.AbstractMetrics;
 
 import java.util.*;
 
@@ -31,6 +31,7 @@ import java.util.*;
  * @author Vasilis Vryniotis <bbriniotis@datumbox.com>
  */
 public class ClusteringMetrics extends AbstractMetrics {
+    private static final long serialVersionUID = 1L;
 
     private double purity = 0.0;
     private double NMI = 0.0; //Normalized Mutual Information: I(Omega,Gama) calculation