+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version 0.8.0-SNAPSHOT - Build 20161218
 
 - Improved Validation:
     - Removed the ValidationMetrics from the Algorithms. Now it is a separate object.
-    - Removed the kFold validation from Algorithms. Now we offer a splitter mechanism.
+    - Removed the kFold validation from Algorithms. Now we offer a new validator mechanism.
     - A single KnowledgeBase implementation is now used.
 - Removed the unnecessary n & d model parameters from all models.